Solve AoC Day 1 (both parts).
I've been reading Martin Fowler's book on refactoring. For some reason,
this inspired me to solve today's problem with a ludicrous number of
classes and with no functions longer than one line. I don't think this
is quite what Fowler was going for.

Needless to say, I will not be doing it this way again.